Create README.md
https://terminhector84.github.io/truzillator/

About Truzillator
Truzillator is a real estate web site, side project, intended to demonstrate the author’s Front -End Developer capabilities. Truzillator leverages Object Oriented Programming’s flexibility to build multiple instances of the same object with shared functionality that is used to manipulate not only the content of each object but also it helps to determine how and when the object is displayed. Each object represents a container with an image and text content that describes a specific property (house, condo, co-op, town house) and it is created dynamically by manipulating the HTML DOM. Truzillator also sports a logo and an Ad Banner created with the HTML5 Canvas API. Other Features include CSS Animations aimed at guiding the user through the steps required to properly display Truzilllator’s housing properties. Media Queries are used to make the site responsive for desktop resolutions it is worth noting, though, that this site is not responsive for mobile.
Skills Used
•	HTML5
•	HTML5 Canvas
•	CSS3
•	CSS3 Animations
•	Media Queries
•	JavaScript
•	Object Oriented JavaScript

Other Credits: Images used with copyright free permission from https://pixabay.com/